Wanderly/backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
python-jose[cryptography]
supabase
python-multipart
passlib[bcrypt]
pydantic[email]
fastapi-mail